+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### New features
 
 - [#3359](https://github.com/clojure-emacs/cider/pull/3359): Add customizable default connection params
+- [#3828](https://github.com/clojure-emacs/cider/issues/3828): Inspector: diff mode.
+- [#3828](https://github.com/clojure-emacs/cider/issues/3828): Inspector: sorting maps by keys.
 
 ### Changes